Move DataToggling logic to a service This might fix problems with data sometimes not being disabled
